refactor: blobs state files
